+++ osmos-core/src/sensor.rs
@@ -1,10 +1,10 @@
 pub struct Sensor {
-    pub range: f32,
-    pub data_list: Vec<f32>,
+    pub range: f64,
+    pub data_list: Vec<f64>,
 }
 
 impl Sensor {
-    pub fn new(range: f32) -> Self {
+    pub fn new(range: f64) -> Self {
         Self {
             range,
             data_list: vec![],
